openMP

+++ tapeMatrix/solver_serial.h
@@ -11,38 +11,38 @@ namespace band_matrix_serial
     {
         struct DecomposeMatrix result;
 
-        size_t n = matrix.n;
-        size_t b = matrix.b;
+        int32_t n = matrix.n;
+        int32_t b = matrix.b;
 
         result.l = (double**)malloc(n * sizeof(double*));
         result.u = (double**)malloc(n * sizeof(double*));
 
-        for (size_t i = 0; i < n; i++)
+        for (int32_t i = 0; i < n; i++)
         {
             result.l[i] = (double*)calloc(n, sizeof(double));
             result.u[i] = (double*)malloc(n * sizeof(double));
         }
 
-        for (size_t i = 0; i < n; i++)
+        for (int32_t i = 0; i < n; i++)
         {
-            for (size_t j = 0; j < n; j++)
+            for (int32_t j = 0; j < n; j++)
             {
                 result.u[i][j] = matrix.A[i][j];
             }
         }
 
-        for (size_t i = 0; i < n; i++)
+        for (int32_t i = 0; i < n; i++)
         {
             result.l[i][i] = 1.0;
         }
 
-        for (size_t k = 0; k < n - 1; ++k)
+        for (int32_t k = 0; k < n - 1; ++k)
         {
-            size_t upper_bound = (k + b + 1 < n) ? (k + b + 1) : n;
-            for (size_t i = k + 1; i < upper_bound; ++i)
+            int32_t upper_bound = (k + b + 1 < n) ? (k + b + 1) : n;
+            for (int32_t i = k + 1; i < upper_bound; ++i)
             {
                 result.l[i][k] = result.u[i][k] / result.u[k][k];
-                for (size_t j = k; j < upper_bound; ++j)
+                for (int32_t j = k; j < upper_bound; ++j)
                 {
                     result.u[i][j] -= result.l[i][k] * result.u[k][j];
                 }
@@ -51,10 +51,10 @@ namespace band_matrix_serial
         return result;
     }
 
-    void reverse_array(double* array, size_t n)
+    void reverse_array(double* array, int32_t n)
     {
         double temp;
-        for (size_t i = 0; i < n / 2; ++i)
+        for (int32_t i = 0; i < n / 2; ++i)
         {
             temp = array[i];
             array[i] = array[n - 1 - i];
@@ -64,13 +64,13 @@ namespace band_matrix_serial
 
     void solve_lu(DecomposeMatrix decompose_matrix, struct Matrix* matrix)
     {
-        size_t n = matrix->n;
+        int32_t n = matrix->n;
         double* y = (double*)malloc(n * sizeof(double));
 
-        for (size_t i = 0; i < n; i++)
+        for (int32_t i = 0; i < n; i++)
         {
             double s = 0.0;
-            for (size_t j = 0; j < i; j++)
+            for (int32_t j = 0; j < i; j++)
             {
                 s += decompose_matrix.l[i][j] * y[j];
             }
@@ -79,10 +79,10 @@ namespace band_matrix_serial
 
         matrix->X = (double*)malloc(n * sizeof(double));
 
-        for (size_t i = n - 1, k = 0; i >= 0; --i, k++)
+        for (int32_t i = n - 1, k = 0; i >= 0; --i, k++)
         {
             double s = 0.0;
-            for (size_t j = n - 1; j > i; --j)
+            for (int32_t j = n - 1; j > i; --j)
             {
                 s += decompose_matrix.u[i][j] * matrix->X[n - j - 1];
             }
